+++ hnote/components/atoms/ParagraphBold.tsx
@@ -3,8 +3,15 @@ import { Text, View, StyleSheet } from 'react-native';
 
 const ParagraphBold = () => (
     <View>
-        <Text>{Children}</Text>
+        <Text style={styles.text}>{Children}</Text>
     </View>
 );
+
+const styles = StyleSheet.create({
+    wrapper: {},
+    text: {
+        fontSize: 700,
+    },
+});
 
 export default ParagraphBold;