+++ hnote/components/organisms/Sidebars/MainNavbar.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import { View } from 'react-native';
+import UserHeading from '../../molecules/UserHeading';
+
+const MainNavbar = () => {
+    return (
+        <View>
+            <UserHeading></UserHeading>
+        </View>
+    );
+};
+
+export default MainNavbar;